Refactor FormStory
