+++ src/components/pages/FormAdditionalInformation.js
@@ -1,5 +1,5 @@
 import React from 'react';
-import PropTypes from 'prop-types'
+import PropTypes from 'prop-types';
 
 import { TextField, Typography } from '@material-ui/core';
 
@@ -36,9 +36,9 @@ function FormAdditionalInformation({ data, setData }) {
 FormAdditionalInformation.propTypes = {
   data: PropTypes.shape({
     title: PropTypes.string.isRequired,
-    tags: PropTypes.arrayOf(PropTypes.string).isRequired
+    tags: PropTypes.arrayOf(PropTypes.string).isRequired,
   }).isRequired,
-  setData: PropTypes.func.isRequired
-}
+  setData: PropTypes.func.isRequired,
+};
 
 export default FormAdditionalInformation;
